+++ src/main/java/org/example/PersonFacade.java
@@ -6,6 +6,7 @@ import org.example.exceptions.NotFoundPersonException;
 import org.example.model.Person;
 import org.example.model.PersonType;
 import org.example.repository.PersonRepository;
+import org.example.utils.FileDeleter;
 import org.example.utils.PersonValidator;
 import org.example.utils.XmlWriter;
 import org.xml.sax.SAXException;
