+++ src/main/java/org/example/utils/FileDeleter.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.utils;
 
 import java.io.File;
